# Faster algorithm for suppress_non_maximum